+++ packages/stylelint-config-ibmdotcom/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/stylelint-config-ibmdotcom@1.11.1-rc.0...@carbon/stylelint-config-ibmdotcom@1.11.1) (2021-06-01)
+
+**Note:** Version bump only for package @carbon/stylelint-config-ibmdotcom
+
+
+
+
+
+## [1.11.1-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/stylelint-config-ibmdotcom@1.11.0...@carbon/stylelint-config-ibmdotcom@1.11.1-rc.0) (2021-05-22)
+
+**Note:** Version bump only for package @carbon/stylelint-config-ibmdotcom
+
+
+
+
+
 # [1.11.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/stylelint-config-ibmdotcom@1.11.0-rc.0...@carbon/stylelint-config-ibmdotcom@1.11.0) (2021-04-21)
 
 **Note:** Version bump only for package @carbon/stylelint-config-ibmdotcom